requirements.txt
ECPy
pyjwt
dataclasses-json